Use name 'rest_step' for function in charge of observing subscriptions
